fix: Corrected an issue in the regex configuration format and final read format
Two issues were fixed here. The first issue was in the yaml configuration files. where the regex lines were intended 1 time to many. The second issue was that before reading the regex value in as a string to a Regex value the newline characters "\n" were not being removed from the string.